Second iteration

+++ src/subject/subject.controller.ts
@@ -7,6 +7,7 @@ class SubjectCtrl {
   private subject
   private days
   private input
+  private homeworkTypes
 
   constructor(public $state,
               public $scope,
@@ -16,7 +17,9 @@ class SubjectCtrl {
               public SubjectService) {
     this.updateSubject()
     this.days = this.CalendarService.getDays()
+    this.homeworkTypes = this.SubjectService.getHomeworkTypes()
     this.setModals()
+    this.resetInput()
   }
 
   updateSubject() {
@@ -25,25 +28,29 @@ class SubjectCtrl {
 
   setModals() {
     let scope = this.$scope
-    this.ModalService.createModal(scope, 'templates/subject-add-class.html', 'addClass')
-    this.ModalService.createModal(scope, 'templates/subject-add-exam.html', 'addExam')
-    this.ModalService.createModal(scope, 'templates/subject-edit-exam.html', 'editExam')
-    this.ModalService.createModal(scope, 'templates/subject-add-homework.html', 'addHomework')
-    this.ModalService.createModal(scope, 'templates/subject-edit-homework.html', 'editHomework')
-    this.ModalService.createModal(scope, 'templates/subject-edit-finalgrade.html', 'editFinalGrade')
+    let modal = this.ModalService
+    modal.createModal(scope, 'templates/subject-add-class.html', 'addClass')
+    modal.createModal(scope, 'templates/subject-add-exam.html', 'addExam')
+    modal.createModal(scope, 'templates/subject-edit-exam.html', 'editExam')
+    modal.createModal(scope, 'templates/subject-add-homework.html', 'addHomework')
+    modal.createModal(scope, 'templates/subject-edit-homework.html', 'editHomework')
+    modal.createModal(scope, 'templates/subject-edit-finalgrade.html', 'editFinalGrade')
+    modal.createModal(scope, 'templates/subject-add-homework-type.html', 'addHomeworkType')
   }
 
   resetInput() {
+    let homeworkTypes = this.getHomeworkTypes()
     this.input = {}
     this.input['addClass'] = { day: '1' }
     this.input['addExam'] = { weight: 1 }
-    this.input['addHomework'] = { weight: 1 }
-    this.input['editFinalGrade'] = { 
+    this.input['addHomework'] = { weight: 1, homeworkType: homeworkTypes[0] }
+    this.input['editFinalGrade'] = {
       examsWeight: this.subject.examsWeight,
-      homeworksWeight: this.subject.homeworksWeight
+      homeworksWeight: this.subject.homeworksWeight ? this.subject.homeworksWeight : {}
     }
     this.input['editExam'] = {}
     this.input['editHomework'] = {}
+    this.input['addHomeworkType'] = {}
   }
 
   showModal(modalName, propType?, propId?) {
@@ -60,6 +67,8 @@ class SubjectCtrl {
       for (let i = 0; i < attr.length; i++) {
         this.input[modalName][attr[i]] = homework[attr[i]]
       }
+      this.input[modalName].dueTime = this.input[modalName].startTime
+      this.input[modalName].homeworkType = homework.homeworkType
     }
     this.ModalService.showModal(modalName)
   }
@@ -67,6 +76,30 @@ class SubjectCtrl {
   closeModal(modalName) {
     this.updateSubject()
     this.ModalService.closeModal(modalName)
+  }
+
+  existsClassWithSameDate(input) {
+    let clazz = this.subject.classes.find(c => {
+      return c.startTime.getHours() == input.startTime.getHours() &&
+             c.startTime.getMinutes() == input.startTime.getMinutes() &&
+             c.endTime.getHours() == input.endTime.getHours() &&
+             c.endTime.getMinutes() == input.endTime.getMinutes() &&
+             c.day == input.date.getDay()
+    })
+    return clazz != null
+  }
+
+  fillExamTime(modalName) {
+    if (this.input[modalName].date instanceof Date) {
+      let examDay = this.input[modalName].date.getDay()
+      let clazz = this.subject.classes.find(c => {
+        return c.day == examDay
+      })
+      if (clazz) {
+        this.input[modalName].startTime = new Date(clazz.startTime)
+        this.input[modalName].endTime = new Date(clazz.endTime)
+      }
+    }
   }
 
   addClass() {
@@ -78,6 +111,30 @@ class SubjectCtrl {
     this.closeModal('addClass')
   }
 
+  addHomeworkType(modalName) {
+    let input = this.input['addHomeworkType']
+    if (!this.SubjectService.validateHomeworkType(input))
+      return
+    this.SubjectService.addHomeworkType(input.type)
+    this.homeworkTypes.push(input.type)
+    this.input[modalName].homeworkType = input.type
+    this.closeModal('addHomeworkType')
+    this.input['addHomeworkType'] = {}
+  }
+
+  checkHomeworkType(modalName) {
+    let homeworkType = this.input[modalName].homeworkType
+    if (homeworkType == '_new') {
+      this.input['addHomeworkType'].modalName = modalName
+      this.ModalService.showModal('addHomeworkType')
+    }
+  }
+
+  closeHomeworkTypeModal(modalName) {
+    this.input[modalName].homeworkType = ''
+    this.closeModal('addHomeworkType')
+  }
+
   addExam() {
     let input = this.input['addExam']
     if (this.SubjectService.validateTime(this.input) == false) return
@@ -86,19 +143,36 @@ class SubjectCtrl {
     this.SubjectService.addSubjectProperty(this.subject.id, "exams", input)
     this.CalendarService.createEvent(input, this.subject, input.date)
     this.closeModal('addExam')
+    if (this.existsClassWithSameDate(input)) {
+      this.CalendarService.changeEventVisibility(this.subject.id, 'classes',
+        input.startTime, input.endTime, true)
+    }
   }
 
   editExam() {
     let input = this.input['editExam']
+    let oldExam = this.subject.exams.find(s => { return s.id == input.id })
     if (this.SubjectService.validateTime(input) == false) return
     this.SubjectService.editSubjectProperty(this.subject.id, "exams", input)
     this.CalendarService.editChildEvent(this.subject.id, input.id, input)
     this.closeModal('editExam')
+    if (this.existsClassWithSameDate(oldExam) &&
+        !this.CalendarService.equals(oldExam.startTime, input.startTime) &&
+        !this.CalendarService.equals(oldExam.endTime, input.endTime)) {
+      this.CalendarService.changeEventVisibility(this.subject.id, 'classes',
+        oldExam.startTime, oldExam.endTime, false)
+    }
+    if (this.existsClassWithSameDate(input)) {
+      this.CalendarService.changeEventVisibility(this.subject.id, 'classes',
+        input.startTime, input.endTime, true)
+    }
   }
 
   addHomework() {
     let input = this.input['addHomework']
-    if (this.SubjectService.validateTime(this.input) == false) return
+    input.startTime = input.dueTime
+    input.endTime = input.dueTime
+    delete input.dueTime
     this.SubjectService.addSubjectProperty(this.subject.id, "homeworks", input)
     this.CalendarService.createEvent(input, this.subject, input.date)
     this.closeModal('addHomework')
@@ -106,7 +180,9 @@ class SubjectCtrl {
 
   editHomework() {
     let input = this.input['editHomework']
-    if (this.SubjectService.validateTime(input) == false) return
+    input.startTime = input.dueTime
+    input.endTime = input.dueTime
+    delete input.dueTime
     this.SubjectService.editSubjectProperty(this.subject.id, "homeworks", input)
     this.CalendarService.editChildEvent(this.subject.id, input.id, input)
     this.closeModal('editHomework')
@@ -146,7 +222,7 @@ class SubjectCtrl {
           this.deleteSubjectProperty(input.id, 'exams')
           this.closeModal('editExam')
         }
-      }) 
+      })
   }
 
   showConfirmHomework() {
@@ -157,7 +233,7 @@ class SubjectCtrl {
           this.deleteSubjectProperty(input.id, 'homeworks')
           this.closeModal('editHomework')
         }
-      }) 
+      })
   }
 
   delete() {
@@ -167,16 +243,34 @@ class SubjectCtrl {
   }
 
   deleteSubjectProperty(propId, propName) {
+    if (propName == "exams") {
+      let exam = this.subject.exams.find(e => { return e.id == propId })
+      if (this.existsClassWithSameDate(exam)) {
+        this.CalendarService.changeEventVisibility(this.subject.id, "classes",
+            exam.startTime, exam.endTime, false)
+      }
+    }
     this.SubjectService.deleteSubjectProperty(this.subject.id, propName, propId)
     this.CalendarService.deleteChildEvent(this.subject.id, propId, propName)
     this.updateSubject()
+  }
+
+  getMaxWeight() {
+    let types = this.getHomeworkTypesInUse()
+    let input = this.input['editFinalGrade']
+    let maxWeight = 0
+    maxWeight += input.examsWeight ? input.examsWeight : 0
+    types.forEach(t => {
+      maxWeight += input.homeworksWeight[t] ? input.homeworksWeight[t] : 0
+    })
+    return maxWeight
   }
 
   getExamsGrade() {
     let exams = this.subject.exams
     let totalWeight = 0
     let totalScore = 0
-    
+
     if (!this.hasAnyExam()) return 0
     for (let i = 0; i < exams.length; i++) {
       if (exams[i].result != null) {
@@ -187,11 +281,13 @@ class SubjectCtrl {
     return totalScore / totalWeight
   }
 
-  getHomeworksGrade() {
-    let homeworks = this.subject.homeworks
+  getHomeworksGrade(type) {
     let totalWeight = 0
     let totalScore = 0
-    
+    let homeworks = this.subject.homeworks.filter(hw => {
+      return hw.homeworkType == type
+    })
+
     if (!this.hasAnyHomework()) return 0
     for (let i = 0; i < homeworks.length; i++) {
       if (homeworks[i].result != null) {
@@ -203,18 +299,31 @@ class SubjectCtrl {
   }
 
   getFinalGrade() {
-    if (this.hasAnyExam() && this.hasAnyHomework()) {
-      if (this.subject.examsWeight != null && this.subject.homeworksWeight != null) {
-        let examsPart = this.getExamsGrade() * this.subject.examsWeight
-        let homeworksPart = this.getHomeworksGrade() * this.subject.homeworksWeight
-        let totalWeight = this.subject.examsWeight + this.subject.homeworksWeight
-        let result = (examsPart + homeworksPart) / totalWeight
-        return result > 10 ? 10 : result
-      }
-      return null
+    if (this.subject.examsWeight != null && this.subject.homeworksWeight != null) {
+      let examsPart = this.getExamsGrade() * this.subject.examsWeight
+      let homeworksPart = 0
+      this.getHomeworkTypesInUse().forEach(type => {
+        homeworksPart += this.getHomeworksGrade(type) * this.subject.homeworksWeight[type]
+      })
+      let totalWeight = this.getMaxWeight()
+      let result = (examsPart + homeworksPart) / totalWeight
+      return result > 10 ? 10 : result
     }
-    if (this.hasAnyExam()) return this.getExamsGrade()
-    if (this.hasAnyHomework()) return this.getHomeworksGrade()
+    return null
+  }
+
+  getHomeworkTypes() {
+    return this.SubjectService.getHomeworkTypes()
+  }
+
+  getHomeworkTypesInUse() {
+    let types = []
+    this.subject.homeworks.forEach(hw => {
+      if (types.indexOf(hw.homeworkType) < 0) {
+        types.push(hw.homeworkType)
+      }
+    })
+    return types
   }
 }
